Raise an error on write if column is missing or auto-generated
To improve error diagnostics, check that the column is actually a parameter of the prepared statement before attempting to set its value. This means the column is part of the table and is not auto-generated. Raise an exception with the name of the offending column if this is not the case.

Closes #3