Use fully qualified column names in SELECT statements
Currently statements use fully qualified column names only when a
table alias is given. Table aliases are kept in a map, which means
we can't join on the same table twice.

With this change column names are always qualified with the table
alias and tables always have an alias, which might be the table name
itself. This simplifies the handling of aliases, removes the use of
the alias map and allows to join on the same table multiple times by
using different aliases.

Statements are a bit more verbose than before, so we issue statements
with all caps SQL keywords, to make them more readable in logs or
debug outputs.

Closes #24, #26