Make inclusion of join table columns optional
Currently when joining on another table, the columns from that other
table are always include in the result set. This makes the result set
bigger than it might need to be and prevents usages of
`SELECT DISTINCT`.

With this change inclusion of joined tables' columns is off by default.
`Select#join` accepts an extra boolean parameter to control inclusion.

Closes #27