Implement `Count` as a `Select` function
Currently support for SQL `COUNT` is provided via a separate `Count`
statement. This is not ideal as it duplicates some code in `Select`,
has less features, and cannot be combined with `DISTINCT`.

With this change, `count` is no longer a separate statement but
instead a terminating function on `Select`, i.e.:

For example: `Select.distinct(products).where(....).count(connection)`.

Closes #25